// services/graph-rewriting-service/src/types/patternnode.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PatternNodeSchema {
  key: string;
  attributes?: {
    [k: string]: number | string | boolean | null | (number | string | boolean)[];
  };
}